Day Dialogue
Dialogue needs to know when the movement stops, so I combined CharacterMoving with Character.

This also adjust the Character Transform to encompass both the Child Image and the Dialogue Object so Dialogue appears right next to character.

Made a copy of this scene as Raph_Day if merge conflict appears.